♻️ [Refactor] 호버 관련 로직 개선
rename DynamicTexture -> Hover
